Experimental support for match inequalities
With this feature, pattern matching supports numeric inequalities in
addition to the standard equality predicate.  The input bindings
should include a binding for a variable with a name that contains
either "<", ">", "<=", ">=", or "!=" immediately after the leading
"?".  The input pattern can then use that variable.  When matching, a
value X will match that variable only if the binding Y for that
variable satisfies the inequality with X and Y (in that order). In
this case, the output bindings will include a new binding for a
variable with the same name as the inequality variable but without the
actual inequality.

For example, given input bindings {"?<n":10}, pattern {"n":"?<n"}, and
message {"n":3}, the match will succeed with bindings
{"?<n":10,"?n":3}.

See match_test.js/match.md for several examples. (Search for
"inequality".)

Yes, such a feature makes us stare down a slippery slope.

The immediate motivation for this feature was to support timer
fallbacks.  For example, using the Goja interpreter, an action could
establish a binding with a value that's a number representing a future
time in UNIX milliseconds.  Then a branch pattern check for a message
containing the current time which is greater than that number.  When a
machine is loaded, a system could send the machine a "current time"
message, which could advance the machine according to such a branch.
Using this technique, a machine that creates a timer that somehow gets
lots could regain at least some sense of what's going on.

For now at least, the inequalities only work for numeric values.  We
might support strings later.