Handle nil bindings returned by actions; closes #44
The bug report (see #44) that motivated these changes revealed a
problem with a little special case in the code.  Nil bindings were
never given to branches for consideration, which makes sense on its
own because pattern matching cannot (by specification and
implementation) succeed when the given bindings are nil (as opposed to
being empty).  However, this behavior meant that a pattern-less branch
was not followed.  A good question about what we want, and the answer
with this commit is that we want nil bindings returned by an action
(as opposed to a guard) to be equivalent to empty bindings, which
enables default branches to be traversed.

This bug report leads to a deeper issue related to nodes with actions
that, in some circumstances, do not follow any outgoing branch.  This
commit includes comments about this situation.  The short story is
that an older iteration of Sheens prevented this situation via a
richer type system for specs.  This structure was relaxed, and one
consequence is the possibility of a class of runtime
machine/spec-level error, which is what is implemented here.